refactor(voice): extract wireDurableChannels into live-agent-runtime.ts (step 5a-2)
The LiveAgentRuntime's first real content: the durable-work-channel wiring
(context drops, ambient note-viewing with the GOODBYE-trigger metadata-only
guard, task results with the 1.5s/3s retry + Discord-DM/Cartesia stuck-
session fallback) moved VERBATIM from voice-agent main(). This is the
piece every live adapter needs identically — web voice today, phone in
5b-2, MatrixRTC later. Adapter-specifics (Cartesia key + synth) arrive
via DurableChannelOptions.

Equivalence: 15/15 template literals identical (the one diff is the
variable rename in a log-line interpolation — same runtime values);
behavior anchors 11/11; tsc clean. Media frames never touch this module
(three-channel rule).
